`leaflet-fullscreen\fullscreen-map.js` moved to dedicated submodule directory from `script\leaflet-fullscreen.js`
