+++ public/tutorial/accessibility/script/fullscreen.js
@@ -1,7 +1,7 @@
 import { attributionOsm, urlTemplateOsm } from '../../../script/base-layers.js'
 import { marker } from '../../../leaflet-adapter/marker.js'
 import { tileLayer } from '../../../leaflet-adapter/tile-layer/tile-layer.js'
-import { fullscreenMap } from '../../../script/leaflet-fullscreen.js'
+import { fullscreenMap } from '../../../leaflet-fullscreen/fullscreen-map.js'
 
 const altText = 'Kyiv',
   map = fullscreenMap({
